+++ Projects/UOContent/Mobiles/AI/BaseAI/BaseAI.cs
@@ -59,7 +59,7 @@ public abstract class BaseAI
 {
     protected ActionType m_Action;
     public readonly BaseCreature m_Mobile;
-    private long m_NextDetectHidden;
+    public long m_NextDetectHidden;
     private long m_NextStopGuard;
     protected PathFollower m_Path;
     public Timer m_Timer;
@@ -3043,127 +3043,6 @@ public abstract class BaseAI
             {
                 bc.ControlOrder = _order;
             }
-        }
-    }
-
-    private sealed class AITimer : Timer
-    {
-        private readonly BaseAI m_Owner;
-        private int _detectHiddenMinDelay;
-        private int _detectHiddenMaxDelay;
-
-        public AITimer(BaseAI owner) : base(
-            TimeSpan.Zero, TimeSpan.FromMilliseconds(GetBaseInterval(owner)))
-        {
-            m_Owner = owner;
-            m_Owner.m_NextDetectHidden = Core.TickCount;
-        }
-
-        private static double GetBaseInterval(BaseAI owner)
-        {
-            if (owner.m_Mobile.Controlled && owner.m_Mobile.ControlOrder == OrderType.Follow 
-                && owner.m_Mobile.Combatant != owner.m_Mobile.ControlMaster)
-            {
-                return owner.m_Mobile.CurrentSpeed * 500;
-            }
-            else
-            {
-                return owner.m_Mobile.CurrentSpeed * 1000;
-            }
-        }
-
-        protected override void OnTick()
-        {
-            if (ShouldStop())
-            {
-                Stop();
-                return;
-            }
-
-            var newInterval = TimeSpan.FromMilliseconds(GetBaseInterval(m_Owner));
-
-            if (Interval != newInterval)
-            {
-                Interval = newInterval;
-            }
-        
-            m_Owner.m_Mobile.OnThink();
-        
-            if (ShouldStop())
-            {
-                Stop();
-                return;
-            }
-        
-            HandleBardEffects();
-        
-            if (m_Owner.m_Mobile.Controlled ? !m_Owner.Obey() : !m_Owner.Think())
-            {
-                Stop();
-                return;
-            }
-        
-            HandleDetectHidden();
-        }
-
-        private bool ShouldStop()
-        {
-            if (m_Owner.m_Mobile.Deleted)
-            {
-                return true;
-            }
-
-            if (m_Owner.m_Mobile.Map == null || m_Owner.m_Mobile.Map == Map.Internal)
-            {
-                m_Owner.Deactivate();
-                return true;
-            }
-
-            if (m_Owner.m_Mobile.PlayerRangeSensitive && 
-                !m_Owner.m_Mobile.Map.GetSector(m_Owner.m_Mobile.Location).Active)
-            {
-                m_Owner.Deactivate();
-                return true;
-            }
-
-            return false;
-        }
-
-        private void HandleBardEffects()
-        {
-            if (m_Owner.m_Mobile.BardPacified)
-            {
-                m_Owner.DoBardPacified();
-            }
-            else if (m_Owner.m_Mobile.BardProvoked)
-            {
-                m_Owner.DoBardProvoked();
-            }
-        }
-        
-        private void CacheDetectHiddenDelays()
-        {
-            var delay = Math.Min(30000 / m_Owner.m_Mobile.Int, 120);
-            _detectHiddenMinDelay = delay * 900;  // 26s to 108s
-            _detectHiddenMaxDelay = delay * 1100; // 32s to 132s
-        }
-
-        private void HandleDetectHidden()
-        {
-            if (!m_Owner.CanDetectHidden || Core.TickCount - m_Owner.m_NextDetectHidden < 0)
-            {
-                return;
-            }
-
-            m_Owner.DetectHidden();
-
-            if (_detectHiddenMinDelay == 0 || _detectHiddenMaxDelay == 0)
-            {
-                CacheDetectHiddenDelays();
-            }
-
-            m_Owner.m_NextDetectHidden = Core.TickCount + 
-                Utility.RandomMinMax(_detectHiddenMinDelay, _detectHiddenMaxDelay);
         }
     }
